requirements.txt
PySide6>=6.5.0
Pillow>=8.0.0
qdarkstyle>=3.0.0
opencv-python>=4.8.0
numpy>=1.21.0
vpk>=1.3.4
pyinstaller>=5.13.0

